# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and account operations
  - name: subscription
    description: Subscription access state (trial / Pro / expired)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      description: Creates a user with a hashed password and starts email verification. No session is created until the email is verified.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created; email verification pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      description: Verifies credentials and creates a server-side session cookie. Requires a verified email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      description: Destroys the server session and clears the session cookie.
      responses:
        "204":
          description: Signed out

  /auth/me:
    get:
      operationId: me
      tags: [auth]
      summary: Current user
      description: Resolves the signed-in user from the session cookie.
      responses:
        "200":
          description: The signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify an email address
      description: Consumes a single-use verification token and marks the account verified.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          description: Invalid, expired, or already-used token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend the verification email
      description: Issues a new verification token when the account exists and is unverified. The response never reveals whether the account exists. Rate limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryStatusResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset
      description: Issues a single-use reset token when the account exists. The response never reveals whether the account exists. Rate limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryStatusResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password
      description: Consumes a single-use reset token, updates the password, and invalidates all existing sessions for the account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid input, or invalid/expired/used token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscription:
    get:
      operationId: getSubscription
      tags: [subscription]
      summary: Current subscription access state
      description: |-
        Server-derived subscription state for the authenticated user. The
        backend is the single source of truth; the client never computes or
        stores access. States: trialing (3-day trial active), pro (paid
        access active), expired (trial over, no paid access — data intact,
        Pro capabilities blocked).
      responses:
        "200":
          description: The current access state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessResponse"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pro/insights:
    get:
      operationId: getProInsights
      tags: [subscription]
      summary: Example Pro capability (foundation proof)
      description: |-
        A harmless Pro-gated capability that proves the server-side Pro
        authorization guard end-to-end. Real Pro features replace this in a
        later phase. Requires a valid session, a verified email, and Pro
        access (paid or trialing — access policy is decided server-side).
      responses:
        "200":
          description: Capability wired and accessible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProCapabilityResponse"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified, or Pro access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /billing/config:
    get:
      operationId: getBillingConfig
      tags: [billing]
      summary: Frontend-safe billing configuration
      description: |-
        Paddle configuration for the browser: the public client token,
        environment, real price IDs (from env config or live catalog
        discovery), and an honest checkout-availability verdict. Contains no
        secrets — the API key never leaves the server.
      responses:
        "200":
          description: Billing configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingConfigResponse"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /billing/portal-session:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Create a Paddle customer portal session
      description: |-
        Starts a Paddle customer portal session for the signed-in user's
        linked billing customer so they can manage their subscription. Requires
        a verified session and Pro access; 409 when no billing profile is
        linked yet, 502 when the provider is temporarily unavailable.
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPortalSessionResponse"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified, or Pro access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No billing customer linked to this account yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Billing provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 200
      required:
        - name
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - email
        - password

    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - token

    EmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
      required:
        - email

    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 200
        password:
          type: string
          minLength: 8
          maxLength: 200
      required:
        - token
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
      required:
        - id
        - name
        - email
        - emailVerified

    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - user

    RegisterResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        # True only when a provider is configured; the UI must not claim an
        # email was sent when this is false.
        emailDelivered:
          type: boolean
        verificationRequired:
          type: boolean
      required:
        - user
        - emailDelivered
        - verificationRequired

    LoginResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - user

    DeliveryStatusResponse:
      type: object
      properties:
        status:
          type: string
        # Honest delivery state: reflects provider configuration only, never
        # whether the addressed account exists.
        emailDelivered:
          type: boolean
      required:
        - status
        - emailDelivered

    MessageResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    SubscriptionPlan:
      type: object
      description: The active paid plan; null until a real subscription exists.
      properties:
        interval:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [active, canceled]
        cancelAtPeriodEnd:
          type: boolean
      required:
        - interval
        - status
        - cancelAtPeriodEnd

    Subscription:
      type: object
      properties:
        state:
          type: string
          enum: [trialing, pro, expired]
          description: Server-derived access state. hasProAccess is true only for "pro".
        hasProAccess:
          type: boolean
        trialDaysRemaining:
          type: integer
          description: Whole days left in the trial; 0 on the final day and when expired.
        trialStartedAt:
          type: string
          format: date-time
        trialEndsAt:
          type: string
          format: date-time
        trialLengthDays:
          type: integer
        plan:
          $ref: "#/components/schemas/SubscriptionPlan"
        trialEndedAgoMs:
          type: integer
          format: int64
          description: Milliseconds since the trial ended; null unless state is "expired".
      required:
        - state
        - hasProAccess
        - trialDaysRemaining
        - trialStartedAt
        - trialEndsAt
        - trialLengthDays
        - plan
        - trialEndedAgoMs

    SubscriptionAccessResponse:
      type: object
      properties:
        subscription:
          $ref: "#/components/schemas/Subscription"
      required:
        - subscription

    ProCapabilityResponse:
      type: object
      properties:
        status:
          type: string
        capability:
          type: string
        message:
          type: string
      required:
        - status
        - capability
        - message

    BillingConfigResponse:
      type: object
      properties:
        billing:
          type: object
          properties:
            provider:
              type: string
              enum: [paddle]
            environment:
              type: string
              enum: [sandbox, live]
            clientToken:
              type: string
              nullable: true
              description: Paddle client-side token (safe for the browser); null when unset.
            prices:
              type: object
              properties:
                monthly:
                  type: string
                  nullable: true
                yearly:
                  type: string
                  nullable: true
            checkoutAvailable:
              type: boolean
            reason:
              type: string
              nullable: true
              description: Machine-readable reason when checkoutAvailable is false.
          required: [provider, environment, prices, checkoutAvailable, reason]
      required: [billing]

    BillingPortalSessionResponse:
      type: object
      properties:
        status:
          type: string
        url:
          type: string
          description: The Paddle customer portal URL to open.
      required:
        - status
        - url