# lib/api-zod/src/generated/types/billingConfigResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { BillingConfigResponseBilling } from './billingConfigResponseBilling';

export interface BillingConfigResponse {
  billing: BillingConfigResponseBilling;
}