# lib/api-zod/src/generated/types/billingPortalSessionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface BillingPortalSessionResponse {
  status: string;
  /** The Paddle customer portal URL to open. */
  url: string;
}