# lib/api-zod/src/generated/types/subscriptionAccessResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Subscription } from './subscription';

export interface SubscriptionAccessResponse {
  subscription: Subscription;
}