# lib/api-zod/src/generated/types/subscriptionPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { SubscriptionPlanInterval } from './subscriptionPlanInterval';
import type { SubscriptionPlanStatus } from './subscriptionPlanStatus';

/**
 * The active paid plan; null until a real subscription exists.
 */
export interface SubscriptionPlan {
  interval: SubscriptionPlanInterval;
  status: SubscriptionPlanStatus;
  cancelAtPeriodEnd: boolean;
}